pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "routeiq"
version = "2.4.0"
description = "Smart LLM router — auto-selects the optimal model for each task"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "RouteIQ Contributors"},
]
keywords = [
    "llm", "router", "ai", "openai", "anthropic", "claude",
    "gemini", "ollama", "proxy", "cost-optimization",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.109.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=2.2.0",
]
google = [
    "google-genai>=1.0.0",
]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
all = [
    "routeiq[embeddings,google,telemetry]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/osippos/routeiq"
Repository = "https://github.com/osippos/routeiq"
Issues = "https://github.com/osippos/routeiq/issues"
Changelog = "https://github.com/osippos/routeiq/blob/main/CHANGELOG.md"

[project.scripts]
routeiq = "cli:main"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]
